--- email/18-ISbo-2a/logging_config.conf
@@ -1,4 +1,0 @@
-[DEFAULT]
-level=DEBUG
-format=%(asctime)s - %(name)s - %(levelname)s - %(message)s
-file=product_logs.log
